Bug fix: aus.content_size from INT UNSIGNED to BIGINT UNSIGNED
git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@10085 4f837ed2-42f5-46e7-a7a5-fa17313484d4

+++ scripts/slurp/aus.sql
@@ -58,7 +58,7 @@ CREATE TABLE IF NOT EXISTS %s (
     last_completed_poll DATETIME,
     
     /* The AU's content size */
-    content_size INT UNSIGNED,
+    content_size BIGINT UNSIGNED,
     
     /* The AU's disk usage in MB */
     disk_usage FLOAT UNSIGNED
